+++ Tests/Unit/Cache/PagesCacheServiceTest.php
@@ -17,16 +17,23 @@ use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\MockObject\Stub;
 use PHPUnit\Framework\TestCase;
+use Psr\Http\Message\MessageInterface;
+use Psr\Http\Message\RequestInterface;
+use Psr\Http\Message\ServerRequestInterface;
+use Psr\Http\Message\StreamInterface;
+use Psr\Http\Message\UriInterface;
 use TYPO3\CMS\Core\Cache\Frontend\FrontendInterface;
+use TYPO3\CMS\Core\Information\Typo3Version;
+use TYPO3\CMS\Frontend\Cache\CacheInstruction;
 use TYPO3\CMS\Frontend\Controller\TypoScriptFrontendController;
 
 #[CoversClass(PagesCacheService::class)]
 final class PagesCacheServiceTest extends TestCase
 {
     private PagesCacheService $subject;
-    private FrontendInterface&MockObject $cacheFrontendMock;
-
-    private TypoScriptFrontendController&Stub $controllerStub;
+    private MockObject $cacheFrontendMock;
+    private Stub $controllerStub;
+    private ServerRequestInterface $request;
 
     protected function setUp(): void
     {
@@ -40,6 +47,14 @@ final class PagesCacheServiceTest extends TestCase
 
         $this->subject = new PagesCacheService($this->cacheFrontendMock);
         $this->subject->setTypoScriptFrontendController($this->controllerStub);
+
+        $this->request = $this->buildRequest();
+        $GLOBALS['TYPO3_REQUEST'] = $this->request;
+    }
+
+    protected function tearDown(): void
+    {
+        unset($GLOBALS['TYPO3_REQUEST']);
     }
 
     #[Test]
@@ -63,6 +78,42 @@ final class PagesCacheServiceTest extends TestCase
             ->willReturn('some markup');
 
         self::assertSame('some markup', $this->subject->getMarkupFromCache());
+    }
+
+    #[Test]
+    public function getMarkupFromCacheReturnsNullIfCachingIsDisabledViaCacheInstruction(): void
+    {
+        if ((new Typo3Version())->getMajorVersion() < 13) {
+            self::markTestSkipped('Only TYPO3 v13+');
+        }
+
+        $cacheInstruction = new CacheInstruction();
+        $cacheInstruction->disableCache('Test');
+        $GLOBALS['TYPO3_REQUEST'] = $this->request->withAttribute('frontend.cache.instruction', $cacheInstruction);
+
+        $this->cacheFrontendMock
+            ->expects(self::never())
+            ->method('get')
+            ->with('some-hash-tx-schema');
+
+        self::assertNull($this->subject->getMarkupFromCache());
+    }
+
+    #[Test]
+    public function getMarkupFromCacheReturnsNullIfCachingIsDisabledViaTSFE(): void
+    {
+        if ((new Typo3Version())->getMajorVersion() >= 13) {
+            self::markTestSkipped('Only TYPO3 v11/v12');
+        }
+
+        $this->controllerStub->no_cache = true;
+
+        $this->cacheFrontendMock
+            ->expects(self::never())
+            ->method('get')
+            ->with('some-hash-tx-schema');
+
+        self::assertNull($this->subject->getMarkupFromCache());
     }
 
     #[Test]
@@ -93,5 +144,210 @@ final class PagesCacheServiceTest extends TestCase
             ->with('some-hash-tx-schema', 'markup to store', ['pageId_42', 'some_tag_1', 'some_tag_2'], 86400);
 
         $this->subject->storeMarkupInCache('markup to store');
+    }
+
+    #[Test]
+    public function storeMarkupInCacheDoesNotSetMarkupIfCachingIsDisabledViaCacheInstruction(): void
+    {
+        if ((new Typo3Version())->getMajorVersion() < 13) {
+            self::markTestSkipped('Only TYPO3 v13+');
+        }
+
+        $cacheInstruction = new CacheInstruction();
+        $cacheInstruction->disableCache('Test');
+        $GLOBALS['TYPO3_REQUEST'] = $this->request->withAttribute('frontend.cache.instruction', $cacheInstruction);
+
+        $this->cacheFrontendMock
+            ->expects(self::never())
+            ->method('set');
+
+        $this->subject->storeMarkupInCache('markup to store');
+    }
+
+    #[Test]
+    public function storeMarkupInCacheDoesNotSetMarkupIfCachingIsDisabledViaTSFE(): void
+    {
+        if ((new Typo3Version())->getMajorVersion() >= 13) {
+            self::markTestSkipped('Only TYPO3 v11/v12');
+        }
+
+        $this->controllerStub->no_cache = true;
+
+        $this->cacheFrontendMock
+            ->expects(self::never())
+            ->method('set');
+
+        $this->subject->storeMarkupInCache('markup to store');
+    }
+
+    private function buildRequest(): ServerRequestInterface
+    {
+        return new class implements ServerRequestInterface {
+            private ?CacheInstruction $cacheInstruction;
+
+            public function __construct()
+            {
+                $this->cacheInstruction = \class_exists(CacheInstruction::class) ? new CacheInstruction() : null;
+            }
+
+            public function getProtocolVersion(): string
+            {
+                throw new \Exception('unused');
+            }
+
+            public function withProtocolVersion(string $version): MessageInterface
+            {
+                throw new \Exception('unused');
+            }
+
+            public function getHeaders(): array
+            {
+                throw new \Exception('unused');
+            }
+
+            public function hasHeader(string $name): bool
+            {
+                throw new \Exception('unused');
+            }
+
+            public function getHeader(string $name): array
+            {
+                throw new \Exception('unused');
+            }
+
+            public function getHeaderLine(string $name): string
+            {
+                throw new \Exception('unused');
+            }
+
+            public function withHeader(string $name, $value): MessageInterface
+            {
+                throw new \Exception('unused');
+            }
+
+            public function withAddedHeader(string $name, $value): MessageInterface
+            {
+                throw new \Exception('unused');
+            }
+
+            public function withoutHeader(string $name): MessageInterface
+            {
+                throw new \Exception('unused');
+            }
+
+            public function getBody(): StreamInterface
+            {
+                throw new \Exception('unused');
+            }
+
+            public function withBody(StreamInterface $body): MessageInterface
+            {
+                throw new \Exception('unused');
+            }
+
+            public function getRequestTarget(): string
+            {
+                throw new \Exception('unused');
+            }
+
+            public function withRequestTarget(string $requestTarget): RequestInterface
+            {
+                throw new \Exception('unused');
+            }
+
+            public function getMethod(): string
+            {
+                throw new \Exception('unused');
+            }
+
+            public function withMethod(string $method): RequestInterface
+            {
+                throw new \Exception('unused');
+            }
+
+            public function getUri(): UriInterface
+            {
+                throw new \Exception('unused');
+            }
+
+            public function withUri(UriInterface $uri, bool $preserveHost = false): RequestInterface
+            {
+                throw new \Exception('unused');
+            }
+
+            public function getServerParams(): array
+            {
+                throw new \Exception('unused');
+            }
+
+            public function getCookieParams(): array
+            {
+                throw new \Exception('unused');
+            }
+
+            public function withCookieParams(array $cookies): ServerRequestInterface
+            {
+                throw new \Exception('unused');
+            }
+
+            public function getQueryParams(): array
+            {
+                throw new \Exception('unused');
+            }
+
+            public function withQueryParams(array $query): ServerRequestInterface
+            {
+                throw new \Exception('unused');
+            }
+
+            public function getUploadedFiles(): array
+            {
+                throw new \Exception('unused');
+            }
+
+            public function withUploadedFiles(array $uploadedFiles): ServerRequestInterface
+            {
+                throw new \Exception('unused');
+            }
+
+            public function getParsedBody(): void
+            {
+                throw new \Exception('unused');
+            }
+
+            public function withParsedBody($data): ServerRequestInterface
+            {
+                throw new \Exception('unused');
+            }
+
+            public function getAttributes(): array
+            {
+                throw new \Exception('unused');
+            }
+
+            public function getAttribute(string $name, $default = null)
+            {
+                if ($name === 'frontend.cache.instruction') {
+                    return $this->cacheInstruction;
+                }
+
+                throw new \Exception('name not supported');
+            }
+
+            public function withAttribute(string $name, $value): ServerRequestInterface
+            {
+                if ($name === 'frontend.cache.instruction') {
+                    $this->cacheInstruction = $value;
+                    return $this;
+                }
+
+                throw new \Exception('name not supported');
+            }
+
+            public function withoutAttribute(string $name): ServerRequestInterface
+            {
+                throw new \Exception('unused');
+            }
+        };
     }
 }